backend/requirements.txt
FastAPI>=0.109.0
uvicorn[standard]>=0.27.0
Pydantic>=2.0
python-dotenv>=1.0.0
python-dateutil==2.8.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
openai==1.12.0
httpx>=0.24.0,<0.28.0
google-generativeai==0.3.2
hypothesis==6.98.0
pytest==7.4.4
pytest-asyncio==0.23.4
